requirements.txt
numpy>=1.20.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
joblib>=1.0.0
Pillow>=8.0.0
seaborn>=0.13.0
pandas>=1.2.0
jupyter>=1.0.0
notebook>=6.0.0
kaggle>=1.5.0
requests>=2.25.0
pathlib2>=2.3.0 